requirements.txt
anthropic>=0.30.0
openai>=1.0.0
datasets>=2.14.0
python-dotenv>=1.0.0
